+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Change log
   obviously ``False``. This is mostly useful when adding
   ``django-imagefield`` to a project which already has images (which may
   not be processible by Pillow).
+- Fixed the image verification to accept CMYK images again.
 
 
 `0.8`_ (2019-06-21)